HW-3: ProgressBar 메소드 클래스 위치 변경
